requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.4
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-bootstrap3==22.1
django-extensions==3.2.1
django-import-export==3.0.2
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
executing==1.2.0
idna==3.4
ipython==8.6.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.32
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stack-data==0.6.1
tablib==3.3.0
traitlets==5.5.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
